Add files via upload
This project outlines the process of migrating a company's infrastructure to the cloud. It begins by identifying potential issues that may arise during the migration, such as problems with computing, networking, and storage. The project then describes a possible cloud architecture that can be used to address these issues. The architecture is divided into two stages: an on-premises data center, and a suggested cloud architecture. The on-premises data center section discusses the network, computation, and security aspects of this stage. The suggested cloud architecture section describes the architecture itself, as well as the features that it offers and how they align with the company's business objectives. The project also includes information on moving different types of servers to a deployment model such as IaaS or PaaS, and a desktop approach. The project concludes with recommendations on how the company can use the cloud effectively.