Add files via upload
This project is a cloud proposal and design for a web and Internet of Things (IoT) architecture. It includes a suggested architecture for the cloud, a threat and risk assessment report for services and data in the cloud, a proposed IoT design security controls to protect intellectual property, data sovereignty and ensuring compliance with the Code of Practice for IoT security, and a Business Continuity Plan for cloud services to ensure resilience. The project is meant to showcase my knowledge and understanding of the process of creating a secure and functional cloud architecture and Business Continuity Plan for web and IoT services, that I can add to my portfolio.